tags now included in BaseModel
