+++ src/main/java/io/github/infolis/model/TextualReference.java
@@ -4,9 +4,7 @@ import io.github.infolis.util.SerializationUtils;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
@@ -43,7 +41,6 @@ public class TextualReference extends BaseModel {
 	private String			textFile;
 	private String			pattern;
 	private String			mentionsReference;
-	private Set<String> tags = new HashSet<>();
 
 
     public TextualReference() {
@@ -72,14 +69,6 @@ public class TextualReference extends BaseModel {
         this.textFile = textFile;
         this.pattern = pattern;
         this.mentionsReference = mentionsReference;
-    }
-    
-    public void setTags(Set<String> tags) {
-    	this.tags = tags;
-    }
-    
-    public Set<String> getTags() {
-    	return this.tags;
     }
 
 	public String toXML() {
